+++ src/coloc.cpp
@@ -1,0 +1,8 @@
+#include "stdafx.h"
+
+
+int main()
+{
+    return 0;
+}
+
